requirements.txt
discord.py==2.2.3
matplotlib==3.5.1
pandas==1.4.2
python-dotenv==1.1.0
yfinance==0.2.54
